This is a project which was originally in a larger project as a subproject but i have correctly reorganised it neatly and tidied it up
I used a tool I created to automatically update project references stuff like that. It's now building nicely. Not all dependencies are in here. Some are in shared_public which is a generic repository containing sub-repositories for various projects which are not top level or important enough to have their own github repository. Shared infrastructure basically. But this stuff is deemed more echat specific. At least for now but subject to change. I have carefully picked through it to ensure nothing is committed that shouldnt be. This stuff should all be able to be public.