Almost done removing MessageTypes library which has project specific information but is referenced at a low level which is wrong
Have known for sometime to strip it out and the infrastructure that exports message parsing wrappers to javascript and c++ from my anotated C# message, request, response classes is now able to handle this nicely. I parse it a type in the assembly for each set of messages i want to include and it takes it from there.